Refactor: widgets into `FinalizeMotionAndClaimWidget`
