--- src/modules/dashboard/components/ActionsPage/ClaimStakesWidget/ClaimStakesWidget.css.d.ts
@@ -1,5 +1,0 @@
-export const main: string;
-export const item: string;
-export const title: string;
-export const label: string;
-export const value: string;
