--- src/modules/dashboard/components/ActionsPage/ClaimStakesWidget/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ClaimStakesWidget';
