+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionAndClaimWidget/FinalizeMotionAndClaimWidget.css.d.ts
@@ -3,6 +3,7 @@ export const stateMarginBottom: string;
 export const item: string;
 export const itemWithForcedBorder: string;
 export const label: string;
+export const title: string;
 export const value: string;
 export const help: string;
 export const tooltip: string;
