+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionAndClaimWidget/VoteResults/VoteResults.css.d.ts
