+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionAndClaimWidget/VoteResults/VoteResultsItem.css
