+++ src/modules/dashboard/components/ActionsPage/FinalizeMotionAndClaimWidget/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FinalizeMotionAndClaimWidget';
