--- src/modules/dashboard/components/ActionsPage/FinalizeMotionWidget/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FinalizeMotionWidget';
